Replaced TcpForwarding example with TcpFastPathForwarding example
