--- Tcp/TcpFastPathForwarding/.gitattributes
@@ -1,1 +1,0 @@
-*.max filter=lfs diff=lfs merge=lfs -text
