--- Tcp/TcpFastPathForwarding/bitstream/src/com/maxeler/examples/tcpfastpathforwarding/TcpFastPathForwardingKernel.maxj
@@ -1,163 +1,0 @@
-package com.maxeler.examples.tcpfastpathforwarding;
-
-import com.maxeler.examples.tcpfastpathforwarding.TcpFastPathForwardingTypes.DecisionRomStructType;
-import com.maxeler.examples.tcpfastpathforwarding.TcpFastPathForwardingTypes.EventsStructType;
-import com.maxeler.maxcompiler.v2.kernelcompiler.Kernel;
-import com.maxeler.maxcompiler.v2.kernelcompiler.KernelParameters;
-import com.maxeler.maxcompiler.v2.kernelcompiler.stdlib.Reductions;
-import com.maxeler.maxcompiler.v2.kernelcompiler.stdlib.core.IO.NonBlockingInput;
-import com.maxeler.maxcompiler.v2.kernelcompiler.stdlib.memory.Memory;
-import com.maxeler.maxcompiler.v2.kernelcompiler.types.base.DFEType;
-import com.maxeler.maxcompiler.v2.kernelcompiler.types.base.DFEVar;
-import com.maxeler.maxcompiler.v2.kernelcompiler.types.composite.DFEStruct;
-import com.maxeler.tcp.fastpath.types.MaxTcpLinkType;
-import com.maxeler.udp.fastpath.types.UdpMultiRxType;
-import com.maxeler.udp.fastpath.types.UdpType;
-
-public class TcpFastPathForwardingKernel extends Kernel {
-	public static String INPUT = "inputFeed";
-	public static String OUTPUT_CONSUMERS = "toConsumers";
-	public static String OUTPUT_EVENTS_TO_CPU = "events";
-
-	public TcpFastPathForwardingKernel(KernelParameters params, MaxTcpLinkType consumerLinkType, UdpMultiRxType inputFeedLinkType) {
-		super(params);
-		flush.disabled();
-
-		NonBlockingInput<DFEStruct> input = io.nonBlockingInput(INPUT,
-				inputFeedLinkType.getDFEStructType(),
-				inputFeedLinkType.getDFEStructType().getPositionForField(inputFeedLinkType.getEOF()));
-
-		InputMessage inputMessage = processInputFeed(input, inputFeedLinkType);
-		ConsumerDecision decision = makeDecision(inputMessage);
-
-		io.output(OUTPUT_CONSUMERS, consumerLinkType.getDFEStructType(),
-				inputMessage.valid & decision.shouldForward) <== genConsumerOutput(inputMessage, decision, consumerLinkType);
-
-		io.output(OUTPUT_EVENTS_TO_CPU, TcpFastPathForwardingTypes.eventType,
-				inputMessage.isMessageStart) <== genCpuEvent(inputMessage, decision);
-	}
-
-	private DFEStruct genConsumerOutput(InputMessage inputMessage,	ConsumerDecision decision, MaxTcpLinkType consumerLinkType) {
-		DFEStruct consumerLink = consumerLinkType.getDFEStructType().newInstance(this);
-
-		DFEType socketIdType = (DFEType) consumerLinkType.getDFEStructType().getTypeForField(consumerLinkType.getSocketNumber());
-		DFEVar socketId = decision.targetConsumer.cast(socketIdType);
-		
-		consumerLink[consumerLinkType.getData()] <== inputMessage.data;
-		consumerLink[consumerLinkType.getSOF()] <== inputMessage.sof;
-		consumerLink[consumerLinkType.getEOF()] <== inputMessage.eof;
-		consumerLink[consumerLinkType.getMod()] <== inputMessage.mod;
-		consumerLink[consumerLinkType.getSocketNumber()] <== socketId;
-
-		return consumerLink;
-	}
-
-	private DFEStruct genCpuEvent(InputMessage inputMessage, ConsumerDecision decision) {
-		DFEStruct eventToCpu = TcpFastPathForwardingTypes.eventType.newInstance(this);
-		eventToCpu[EventsStructType.ForwardDecision] <== decision.shouldForward ?
-					constant.var((DFEType)eventToCpu[EventsStructType.ForwardDecision].getType(), 1) :
-					constant.var((DFEType)eventToCpu[EventsStructType.ForwardDecision].getType(), 0);
-		eventToCpu[EventsStructType.ConsumerId] <== decision.targetConsumer;
-		eventToCpu[EventsStructType.EventCount] <== inputMessage.messageCount;
-		eventToCpu[EventsStructType.MessageType] <== inputMessage.messageType;
-		eventToCpu[EventsStructType.Padding] <== constant.var((DFEType)eventToCpu[EventsStructType.Padding].getType(), 0);
-
-		return eventToCpu;
-	}
-
-	private class ConsumerDecision {
-		public DFEVar shouldForward;
-		public DFEVar targetConsumer;
-	}
-
-	private ConsumerDecision makeDecision(InputMessage inputMessage) {
-		ConsumerDecision theDecision = new ConsumerDecision();
-
-		/*
-		 * Decision will be made by lookin up the messageType in a ROM.
-		 */
-
-		Memory<DFEStruct> decisionRom = mem.alloc(TcpFastPathForwardingTypes.decisionRomStructType, 256);
-		decisionRom.mapToCPU("decisionRom");
-		DFEStruct decision = Reductions.streamHold(decisionRom.read(inputMessage.messageType), inputMessage.isMessageStart);
-
-		theDecision.shouldForward = (DFEVar)decision.get(DecisionRomStructType.ShouldForward) !== 0;
-		theDecision.targetConsumer = decision.get(DecisionRomStructType.ConsumerId);
-
-		return theDecision;
-	}
-
-
-	private class InputMessage {
-		public DFEVar messageCount;
-		public DFEVar messageType;
-
-		public DFEVar isMessageStart;
-
-		public DFEVar valid;
-		public DFEVar data;
-		public DFEVar sof;
-		public DFEVar eof;
-		public DFEVar mod;
-	}
-
-	private InputMessage processInputFeed(NonBlockingInput<DFEStruct> input, UdpMultiRxType inputFeedLinkType) {
-		DFEStruct inputStruct = input.data;
-		DFEVar inputValid = input.valid;
-		DFEVar data = inputStruct.get(inputFeedLinkType.getData());
-		DFEVar sof = inputStruct.get(inputFeedLinkType.getSOF());
-		DFEVar eof = inputStruct.get(inputFeedLinkType.getEOF());
-		DFEVar mod = inputStruct.get(inputFeedLinkType.getMod());
-		DFEVar socket = inputStruct.get(UdpType.SOCKET); // TODO
-
-		/*
-		 * Start of message marker
-		 */
-		DFEVar isMessageStart = inputValid & sof;
-
-		debug.simPrintf(isMessageStart, this.getName() + ": Input message on Socket %d\n", socket);
-
-		/*
-		 * Count the number of input packets
-		 */
-		DFEVar messageCount = control.count.makeCounter(
-				control.count.makeParams(64)
-					.withInitValue(0)
-					.withEnable(isMessageStart)
-				).getCount();
-
-
-		/*
-		 * Extract the message type from the packet.
-		 * In this example, the message type is the value of byte at position 'messageTypeBytePosition' of the payload.
-		 */
-		int messageTypeBytePosition = 7;
-
-		DFEVar messageType = Reductions.streamHold(getByte(data, messageTypeBytePosition), isMessageStart);
-
-		InputMessage inputMessage = new InputMessage();
-		inputMessage.valid = inputValid;
-		inputMessage.data = data;
-		inputMessage.sof = sof;
-		inputMessage.eof = eof;
-		inputMessage.mod = mod;
-		inputMessage.messageCount = messageCount;
-		inputMessage.messageType = messageType;
-		inputMessage.isMessageStart = isMessageStart;
-
-
-		return inputMessage;
-	}
-
-	private DFEVar getByte(DFEVar data, int offset) {
-		int bytesPerWord = data.getType().getTotalBytes();
-
-		int wordOffset = offset / bytesPerWord;
-		int byteOffset = offset % bytesPerWord;
-
-		// 8 bits in a byte...
-		return stream.offset(data, wordOffset).slice(byteOffset * 8, 8).cast(dfeUInt(8));
-	}
-
-
-}
