--- Tcp/TcpFastPathForwarding/bitstream/src/com/maxeler/examples/tcpfastpathforwarding/TcpFastPathForwardingParameters.maxj
@@ -1,23 +1,0 @@
-package com.maxeler.examples.tcpfastpathforwarding;
-
-import com.maxeler.maxcompiler.v2.build.EngineParameters;
-import com.maxeler.maxcompiler.v2.managers.DFEModel;
-
-public class TcpFastPathForwardingParameters extends EngineParameters {
-
-	public TcpFastPathForwardingParameters(String[] args) {
-		super(args);
-	}
-
-	@Override
-	protected void declarations() {
-		declareParam(s_parameterDFE, DFEModel.class, DFEModel.ISCA);
-		declareParam(s_parameterMaxFileName, DataType.STRING, "TcpFastPathForwarding");
-		declareParam(s_parameterTarget, Target.class, Target.DFE_SIM);
-	}
-
-//	@Override
-//	public String getBuildName() {
-//		return getMaxFileName() + "_" + getTarget() + "_ss" + getHasStreamStatus() + "_sf" + getStreamFrequency();
-//	}
-}
