--- Tcp/TcpFastPathForwarding/runtime/run_consumers.sh
@@ -1,2 +1,0 @@
-nc -v -l 10000 & 
-nc -v -l 10001 &
