--- Tcp/TcpFastPathForwarding/runtime/run_forwarding.sh
@@ -1,2 +1,0 @@
-./udpforwarding 172.16.50.1 172.16.60.1 225.0.0.37 172.16.60.10 172.16.60.10
-
