--- Tcp/TcpFastPathForwarding/runtime/run_multicast.sh
@@ -1,1 +1,0 @@
-./multicast 172.16.50.10 225.0.0.37
