--- Tcp/TcpForwarding/.gitignore
@@ -1,10 +1,0 @@
-*.o
-*.a
-*.la
-*.so
-*.pyc
-*.pcap
-*.swp
-bin
-.metadata*
-.deps
