+++ Tcp/TcpForwarding/bitstream/build.xml
@@ -10,11 +10,17 @@
 
 	<property name="maxeda"   value="${env.MAXCOMPILERJCP}"/>
 	<property name="maxnet"   value="${env.MAXCOMPILERNETJCP}"/>
-	<property name="maxpowernet"   value="${env.MAXPOWERNETDIR}/bin"/>
+	<property name="maxtcpfp"   value="${env.MAXTCPFPDIR}/lib/MaxTcpFastPath.jar"/>
+	<property name="maxudpfp"   value="${env.MAXUDPFPDIR}/lib/MaxUdpFastPath.jar"/>
+	<property name="maxpower" value="${env.MAXPOWERDIR}/jar/MaxPower.jar"/>
+	<property name="maxpowernet" value="${env.MAXPOWERNETDIR}/jar/MaxPowerNet.jar"/>
 	<path id="bitstreams.local.classpath">
 		<pathelement location="bin"/>
 		<pathelement location="${maxeda}"/>
 		<pathelement location="${maxnet}"/>
+		<pathelement location="${maxtcpfp}"/>
+		<pathelement location="${maxudpfp}"/>
+		<pathelement location="${maxpower}"/>
 		<pathelement location="${maxpowernet}"/>
 	</path>
