+++ Tcp/TcpForwarding/runtime/build.py
@@ -9,17 +9,33 @@ except ImportError, e:
 	print "Couldn't find project-utils modules."
 	sys.exit(1)
 
+dirs = [os.environ['MAXTCPFPDIR'], os.environ['MAXUDPFPDIR']]
+
 MAXFILES = ['TcpForwarding.max']
 sources = ['tcpforwarding.c']
-application = 'tcpforwarding'
-includes = []
-my_cflags = []
-my_ldflags = []
+target = 'tcpforwarding'
+cflags = ['-I%s/include' % dir for dir in dirs]
+ldflags = (['-L%s/lib' % dir for dir in dirs] +
+           ['-lmaxtcp', '-lmaxudp', '-lslicnet'])
 
+simNetConfig = [
+	{
+		'NAME' : 'QSFP_TOP_10G_PORT1',
+		'DFE': '172.16.50.1',
+		'TAP': '172.16.50.10',
+		'NETMASK' : '255.255.255.0'
+	},
+	{
+		'NAME' : 'QSFP_BOT_10G_PORT1',
+		'DFE': '172.16.60.1',
+		'TAP': '172.16.60.10',
+		'NETMASK' : '255.255.255.0'
+	}
+]
 
 b = MaxRuntimeBuilder(maxfiles=MAXFILES)
 s = MaxCompilerSim(dfeModel="ISCA")
-e = Executor(logPrefix="[%s] " % (application))
+e = Executor(logPrefix="[%s] " % (target))
 
 def build():
 	compile()
@@ -27,16 +43,16 @@ def build():
 
 def compile():
 	b.slicCompile()
-	b.compile(sources, extra_cflags=my_cflags)
+	b.compile(sources, extra_cflags=cflags)
 
 def link():
-	b.link(sources, application, extra_ldflags=my_ldflags)
+	b.link(sources, target, extra_ldflags=ldflags)
 
 def clean():
 	b.clean()
 
 def start_sim():
-	s.start()
+	s.start(netConfig=simNetConfig)
 
 def stop_sim():
 	s.stop()
@@ -47,7 +63,12 @@ def restart_sim():
 def run_sim():
 	build()
 	s.start()
-	e.execCommand([ "./" + application, "172.16.50.1", "172.16.60.1", "225.0.0.37", "172.16.60.10", "172.16.60.10"])
+	e.execCommand([ "./" + target,
+		simNetConfig[0]['DFE'],
+		simNetConfig[1]['DFE'],
+		"225.0.0.37",
+		simNetConfig[1]['TAP'],
+		simNetConfig[1]['TAP']])
 	e.wait()
 #	s.stop()
